+++ src/Drivers/check_spo.cpp
@@ -26,6 +26,7 @@
 #include <Input/Input.hpp>
 #include <QMCWaveFunctions/einspline_spo.hpp>
 #include <QMCWaveFunctions/einspline_spo_ref.hpp>
+#include <Drivers/pseudo.hpp>
 #include <Utilities/qmcpack_version.h>
 #include <getopt.h>
