+++ src/Input/Input.hpp
@@ -16,6 +16,5 @@
 
 #include <Input/nio.hpp>
 //#include <Input/graphite.hpp>
-#include <Input/pseudo.hpp>
 
 #endif
